chore(repo): Add `deps` check command
Adds the `deps check` command to validate repo dependencies against a global map of pinned/validated versions. Currently only supports direct dependencies. Example output:

```
$ aft deps check
/Users/nydillon/dev/amplify-dart/packages/amplify_core
Mismatched dependency (uuid):
Expected 3.0.6
Found ^3.0.1

/Users/nydillon/dev/amplify-dart/packages/auth/amplify_auth_cognito_dart
Mismatched dependency (uuid):
Expected 3.0.6
Found ^3.0.0
```

commit-id:5c27d6dd
