+++ pubspec.yaml
@@ -13,6 +13,7 @@ dependencies:
   cli_util: ^0.3.5
   json_annotation: ^4.4.0
   path: any
+  pub_semver: ^2.1.1
   pubspec_parse: ^1.2.0
   smithy:
     hosted: https://pub.dillonnys.com
@@ -22,6 +23,7 @@ dependencies:
     version: ^0.5.0
   stream_transform: ^2.0.0
   yaml: ^3.1.0
+  yaml_edit: ^2.0.3
 
 dev_dependencies:
   amplify_lints: ^1.0.0
