+++ tool/commands/amplify_command.dart
@@ -15,7 +15,15 @@
 import 'dart:io';
 
 import 'package:args/command_runner.dart';
+import 'package:checked_yaml/checked_yaml.dart';
 import 'package:path/path.dart' as p;
+import 'package:stream_transform/stream_transform.dart';
+
+import '../models.dart';
+
+const _ignorePackages = [
+  'synthetic_package',
+];
 
 abstract class AmplifyCommand<T> extends Command<T> {
   /// The root directory of the Amplify Flutter repo.
@@ -33,5 +41,39 @@ abstract class AmplifyCommand<T> extends Command<T> {
       'Root directory not found. Make sure to run this command '
       'from within the Amplify Flutter repo',
     );
+  }();
+
+  /// All packages in the Amplify Flutter repo.
+  late final List<PackageInfo> allPackages = () {
+    final allDirs = rootDir
+        .listSync(recursive: true, followLinks: false)
+        .whereType<Directory>();
+
+    final allPackages = <PackageInfo>[];
+    for (final dir in allDirs) {
+      final pubspec = dir.pubspec;
+      if (pubspec == null) {
+        continue;
+      }
+      if (_ignorePackages.contains(pubspec.name)) {
+        continue;
+      }
+      allPackages.add(PackageInfo(
+        name: pubspec.name,
+        path: dir.path,
+        usesMonoRepo: dir.usesMonoRepo,
+        pubspec: pubspec,
+        flavor: pubspec.flavor,
+      ));
+    }
+    return allPackages..sort();
+  }();
+
+  /// The global dependency configuration for the repo.
+  late final GlobalDependencyConfig globalDependencyConfig = () {
+    final depsFile = File(p.join(rootDir.path, 'deps.yaml'));
+    assert(depsFile.existsSync(), 'Could not find deps.yaml');
+    final depsYaml = depsFile.readAsStringSync();
+    return checkedYamlDecode(depsYaml, GlobalDependencyConfig.fromJson);
   }();
 }
